feat: integrate MarkdownHint directly into MarkdownTextarea for improved usability
